# Optimizations for Matrix4, BoundingBox and Frustum