@giawa 2020-09-02
Very cool!  This looks much nicer.  I support merging this if all tests are looking good.